Refactor REPL
